+++ src/main/java/cal/Calculator.java
@@ -1,4 +1,4 @@
-package cal.utils;
+package cal;
 
 
 import java.util.HashMap;
